pyproject.toml
[tool.poetry]
name = "encdemo"
version = "0.1.0"
description = ""
authors = ["cxzlw <wsc_zds@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pycryptodome = "^3.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

